% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dotplot.pbo}
\alias{dotplot.pbo}
\title{Draws an annotated dot plot of study selection sorted by in-sample selection frequency.}
\usage{
\method{dotplot}{pbo}(p, data = NULL, main,
  xlab = "Sorted Study Number (N)", ylab = "IS Selection Frequency",
  show_config = TRUE, show_grid = TRUE, sel_threshold = 50, ...)
}
\arguments{
\item{p}{a 'pbo' object as returned by pbo()}

\item{data}{should not be used}

\item{main}{plot title, default computed internally, passed to dotplot()}

\item{xlab}{x-axis label with default, passed to dotplot()}

\item{ylab}{y-axis label with default, passed to dotplot()}

\item{show_config}{whether to show the study dimension annotations, default TRUE}

\item{show_grid}{whether to show the grid panel, default TRUE}

\item{sel_threshold}{the minimum IS frequency subsetting threshold, default 50}

\item{...}{other parameters as passed to dotplot()}
}
\description{
Draws an annotated dot plot of study selection sorted by in-sample selection frequency.
}

