% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pbo_logit}
\alias{pbo_logit}
\title{Defunct - draws an annotated histogram of rank logits.
#}
\usage{
pbo_logit(...)
}
\arguments{
\item{...}{all arguments ignored}
}
\description{
Defunct - draws an annotated histogram of rank logits.
#
}
\seealso{
'histogram.pbo'
}
