% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pbo}
\alias{pbo}
\title{Perform probability of backtest overfitting computation via CSCV.}
\usage{
<<<<<<< HEAD
  pbo(M, S = 4, ep = NA, threshold = 0, inf_sub = 6)
=======
pbo(M, S = 4, F = NA, threshold = 0, inf_sub = 6,
  allow_parallel = FALSE)
>>>>>>> Reworked into class with lattice method extensions.  Added parallel computing support.
}
\arguments{
\item{M}{a \eqn{TxN} data frame of returns, where \eqn{T} is the samples per study and \eqn{N} is the number of studies.}

\item{S}{the number of subsets of \eqn{M} for CSCV combinations; must evenly divide \eqn{M}}

<<<<<<< HEAD
  \item{ep}{the function to evaluate a study's performance;
  required}
=======
\item{F}{the function to evaluate a study's performance; required}
>>>>>>> Reworked into class with lattice method extensions.  Added parallel computing support.

\item{threshold}{the performance metric threshold (e.g. 0 for Sharpe, 1 for Omega)}

\item{inf_sub}{infinity substitution value for reasonable plotting}

\item{allow_parallel}{whether to enable parallel processing, default FALSE}
}
\value{
object of class 'pbo' containing list of PBO calculation results and settings
}
\description{
<<<<<<< HEAD
  Perform probability of backtest overfitting computation
  via CSCV.
=======
PBO performs probability of backtest overfitting computation via CSCV.
>>>>>>> Reworked into class with lattice method extensions.  Added parallel computing support.
}

