% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pbo_performance_cases}
\alias{pbo_performance_cases}
\title{Defunct - draws a scatter XY plot showing selected study by CSCV case, with out-of-sample
rank exceeding a given threshold.}
\usage{
pbo_performance_cases(...)
}
\arguments{
\item{...}{all arguments ignored}
}
\description{
Defunct - draws a scatter XY plot showing selected study by CSCV case, with out-of-sample
rank exceeding a given threshold.
}
\seealso{
'xyplot.pbo'
}

