% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pbo_performance_dot}
\alias{pbo_performance_dot}
\title{Defunct - draws an annotated dot plot of study selection sorted by selection frequency.}
\usage{
pbo_performance_dot(...)
}
\arguments{
\item{...}{all arguments ignored}
}
\description{
Defunct - draws an annotated dot plot of study selection sorted by selection frequency.
}
\seealso{
'dotplot.pbo'
}

