% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pbo_performance_pairs}
\alias{pbo_performance_pairs}
\title{Defunct - draws a scatter XY plot of study selection pairs with OOS performance
exceeding a given threshold.}
\usage{
pbo_performance_pairs(...)
}
\arguments{
\item{...}{all arguments ignored}
}
\description{
Defunct - draws a scatter XY plot of study selection pairs with OOS performance
exceeding a given threshold.
}
\seealso{
'xyplot.pbo'
}

