% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pbo_performance_xy}
\alias{pbo_performance_xy}
\title{Defunct - draws an annotated XY plot of study selection sorted by selection frequency.}
\usage{
pbo_performance_xy(...)
}
\arguments{
\item{...}{all arguments ignored}
}
\description{
Defunct - draws an annotated XY plot of study selection sorted by selection frequency.
}
\seealso{
'xyplot.pbo'
}

