% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pbo_show_config}
\alias{pbo_show_config}
\title{Writes grid text to a default predetermined location.}
\usage{
pbo_show_config(p)
}
\arguments{
\item{p}{an object of class \code{pbo} as returned by \code{\link[pbo]{pbo}}.}
}
\description{
Writes grid text to a default predetermined location.
}
\note{
Meant for internal use only.
}

