% Generated by roxygen2 (4.0.0): do not edit by hand
\name{summary.pbo}
\alias{summary.pbo}
\title{Summary of backtest overfitting analysis}
\usage{
\method{summary}{pbo}(object, ...)
}
\arguments{
\item{object}{an object of class \code{pbo} usually acquired by
running \code{\link[pbo]{pbo}}.}

\item{...}{additional parameters ignored}
}
\description{
Provides a fitted summary of the results in the \code{pbo} object.
}
\details{
The summary reports the performance evaluation function name and
its threshold level.  The report then extracts from the PBO object and prints
the probability of backtest overfit \code{p_bo}, the slope of the performance
degration linear regression \code{slope}, the adjusted R-squared of that
regression \code{ar^2}, and the proportion of cases leading to losses
\code{p_loss}.
}
\seealso{
pbo
}
\keyword{PBO}
\keyword{backtest}
\keyword{overfitting}
\keyword{probability}
\keyword{summary}

