% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xyplot.pbo}
\alias{xyplot.pbo}
\title{Draws an annotated plot of performance degradation and probability of loss.}
\usage{
\method{xyplot}{pbo}(p, data = NULL, plotType = "cscv", show_eqn = TRUE,
  show_threshold = TRUE, show_config = TRUE, show_rug = TRUE,
  show_prob = TRUE, show_grid = TRUE, increment = 0.1,
  osr_threshold = 50, sel_threshold = 50, xlab, ylab, main, lwd = 1,
  ylab_left, ylab_right, col_bar, col_line, col_sd1 = "#3366cc",
  col_sd2 = "#339999", lty_sd = c(1, 2, 4), ...)
}
\arguments{
\item{p}{a 'pbo' object as returned by pbo()}

\item{data}{should not be used}

\item{plotType}{one of cscv, degradation, dominance, pairs, ranks or selection}

\item{col_bar}{histogram bar fill color}

\item{col_line}{density plot line color}

\item{col_sd1}{stochastic dominance first line color}

\item{col_sd2}{stochastic dominance second line color}

\item{xlab}{x-axis label, default computed if not provided}

\item{ylab}{y-axis label, default computed if not provided}

\item{main}{plot title, default computed if not provided}

\item{lwd}{line width, default 1, passed to panels and legends}

\item{lty_sd}{line type array for stochastic dominance plot, e.g. c(2,3,5)}

\item{ylab_left}{dominance plot left-hand axis label}

\item{ylab_right}{dominance plot right-hand axis label}

\item{increment}{stochastic dominance distribution generator increment, e.g. 0.1 steps}

\item{osr_threshold}{out-of-sample rank threshold for filtering, default 50}

\item{sel_threshold}{selection frequency threshold for filtering, default 50}

\item{show_eqn}{whether to show the line equation annotation, default TRUE}

\item{show_threshold}{whether to show the probability of loss annotation, default TRUE}

\item{show_config}{whether to show the study dimension annotations, default TRUE}

\item{show_grid}{whether to show the panel grid, default TRUE}

\item{show_prob}{whether to show the probability value in dominance plot, default TRUE}

\item{show_rug}{whether to show scatter rugs near the axes, default TRUE}

\item{...}{other parameters passed to 'xyplot' or its panels}
}
\description{
Draws an annotated plot of performance degradation and probability of loss.
}

