% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice.pbo.R
\name{histogram.pbo}
\alias{histogram.pbo}
\title{PBO rank logits histogram.}
\usage{
\method{histogram}{pbo}(
  x,
  data = NULL,
  show_pbo = TRUE,
  show_regions = TRUE,
  show_config = TRUE,
  col_bar = "#cc99cc",
  col_line = "#3366cc",
  ...
)
}
\arguments{
\item{x}{an object of class \code{pbo} as returned by \code{\link[pbo]{pbo}}.}

\item{data}{should not be used}

\item{show_pbo}{whether to show the PBO value annotation,
default TRUE}

\item{show_regions}{whether to show the overfit region annotations,
default TRUE}

\item{show_config}{whether to show the study dimension annotations,
default TRUE}

\item{col_bar}{histogram bar fill color passed to histogram panel}

\item{col_line}{density plot line color passed to density plot panel}

\item{...}{other parameters passed to \code{\link[lattice]{histogram}},
\code{\link[lattice]{densityplot}}, or \code{\link[lattice]{panel.abline}}.}
}
\description{
Draws an annotated histogram of PBO rank logits.
}
\details{
Uses \pkg{lattice} function \code{\link[lattice]{histogram}},
\code{\link[lattice]{densityplot}}, and
\code{\link[lattice]{panel.abline}} panels together with
class-specific annotations.
}
\seealso{
pbo, dotplot.pbo, xyplot.pbo
}
