% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbo-package.R
\docType{package}
\name{pbo-package}
\alias{pbo-package}
\title{Probability of backtest overfitting.}
\description{
Computes the probability of backtest overfitting
}
\details{
Implements algorithms for computing the probability of
backtest overfitting, performance degradation and probability of loss,
and first- and second-order stochastic dominance,
based on the approach specified in Bailey et al., September 2013.
Provides a collection of pre-configured plots based on \code{lattice} graphics.
}
\references{
See Bailey, David H. and Borwein, Jonathan M. and
Lopez de Prado, Marcos and Zhu, Qiji Jim, The Probability of Back-Test
Overfitting (September 1, 2013). Available at SSRN. See
\url{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2326253}.
}
\author{
Matt Barry \email{mrb@softisms.com}
}
\keyword{CSCV}
\keyword{PBO}
\keyword{backtest}
\keyword{overfitting}
\keyword{probability}
