% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pbo_degradation}
\alias{pbo_degradation}
\title{Defunct - draws an annotated plot of performance degradation and probability of loss.}
\usage{
<<<<<<< HEAD
  pbo_degradation(data,
    main = "OOS Performance Degradation", xlab = "R (OOS)",
    ylab = "R (IS)", col_bar = "#cc99cc",
    col_line = "#3366cc", show_eqn = TRUE,
    show_threshold = TRUE, show_config = TRUE,
    show_rug = TRUE, show_grid = FALSE, ...)
}
\arguments{
  \item{data}{a list as returned by pbo()}

  \item{main}{plot title, default 'OOS Performance
  Degradation'}

  \item{xlab}{X-axis label}

  \item{ylab}{Y-axis label}

  \item{col_bar}{histogram bar fill color}

  \item{col_line}{density plot line color}

  \item{show_eqn}{whether to show the line equation
  annotation, default TRUE}

  \item{show_threshold}{whether to show the probability of
  loss annotation, default TRUE}

  \item{show_config}{whether to show the study dimension
  annotations, default TRUE}

  \item{show_rug}{whether to show the scatter rug near the
  axes}

  \item{show_grid}{whether to show the background grid}

  \item{...}{other parameters passed to 'xyplot' or
  'lmline'}
=======
pbo_degradation(...)
}
\arguments{
\item{...}{all arguments ignored}
>>>>>>> Reworked into class with lattice method extensions.  Added parallel computing support.
}
\description{
Defunct - draws an annotated plot of performance degradation and probability of loss.
}
\seealso{
'xyplot.pbo'
}

