% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pbo_degradation}
\alias{pbo_degradation}
\title{Defunct - draws an annotated plot of performance degradation and probability of loss.}
\usage{
pbo_degradation(...)
}
\arguments{
\item{...}{all arguments ignored}
}
\description{
Defunct - draws an annotated plot of performance degradation and probability of loss.
}
\seealso{
'xyplot.pbo'
}

