% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pbo_dominance}
\alias{pbo_dominance}
\title{Defunct - draws an annotated plot of stochastic dominance.}
\usage{
pbo_dominance(...)
}
\arguments{
\item{...}{all arguments ignored}
}
\description{
Defunct - draws an annotated plot of stochastic dominance.
}
\seealso{
'xyplot.pbo'
}

