% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pbo_show_config}
\alias{pbo_show_config}
\title{Writes grid text to a default location; meant for internal use only.}
\usage{
pbo_show_config(p)
}
\arguments{
\item{p}{an object of class 'pbo' as returned by pbo()}
}
\description{
Writes grid text to a default location; meant for internal use only.
}

